pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "memoryatlas"
version = "0.1.0"
description = "Voice memos to Obsidian notes"
requires-python = ">=3.11,<3.14"
dependencies = [
    "typer>=0.9.0",
    "pyyaml>=6.0",
]

[project.scripts]
atlas = "memoryatlas.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["memoryatlas*"]
